% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accidentsData.R
\docType{data}
\name{accidentsData}
\alias{accidentsData}
\title{Car accident data}
\format{A data frame with 383 observations of 2 columns:
\describe{
  \item{\code{record}}{factor representing the driver status}
  \item{\code{seatBelt}}{factor indicating whether the driver wore a seatbelt}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
accidentsData
}
\description{
A data frame showing the use of seat belt and the driver status 
after a car accident in Greece.
}
