% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirthDeath.R
\docType{data}
\name{BirthDeath}
\alias{BirthDeath}
\title{Number of births and death in Greece}
\format{A data frame with 71 observations of 5 columns:
\describe{
  \item{\code{Year}}{years 1932-2016}
  \item{\code{Deaths}}{number of deaths}
  \item{\code{DeathsRate}}{number of deaths per 1000 citizen}
  \item{\code{Births}}{number of births}
  \item{\code{BirthRate}}{number of births per 1000 citizen}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
BirthDeath
}
\description{
A data frame containing the number of births and deaths 
along with their rates from 1932 to 2016.
}
