% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirthDeath2000.R
\docType{data}
\name{BirthDeath2000}
\alias{BirthDeath2000}
\title{Birth and Deaths before and after 2000}
\format{A data frame with 30 observations of 3 columns:
\describe{
  \item{\code{deaths}}{number of deaths}
  \item{\code{births}}{number of births}
  \item{\code{type}}{factor indicating if the number of births and deaths 
  correspond before 2000 or after 2000}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
BirthDeath2000
}
\description{
A data frame containing samples with the number of births 
and deaths before and after 2000.
}
