% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\docType{data}
\name{births}
\alias{births}
\title{Births in Greece, 1976-1989}
\format{A data frame with 14 observations of 2 columns:
\describe{
  \item{\code{year}}{years from 1976 to 1989}
  \item{\code{rate}}{number of births per 1000 people}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
births
}
\description{
A data frame giving the number of births per 1000 people 
in Greece from 1976 to 1989.
}
