% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuelConsumption.R
\docType{data}
\name{FuelConsumption}
\alias{FuelConsumption}
\title{FuelConsumption}
\format{A data frame with 50 observations of 8 columns:
\describe{
  \item{\code{Geographic.area}}{factor with geographic area of Greece}
  \item{\code{Regions}}{factor with regions of Greece}
  \item{\code{Runits}}{factor with regional units of Greece}
  \item{\code{RON95}}{metric tons of ron 95 consumption}
  \item{\code{RON98_100}}{metric tons of ron 98 consumption}
  \item{\code{DieselBio}}{metric tons of bio diesel consumption}
  \item{\code{LPG}}{metric tons of liquefied petroleum gas consumption}
  \item{\code{DieselC}}{metric tons of heating oil consumption}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
FuelConsumption
}
\description{
A data frame containing the fuel consumption in Greece.
}
