% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggttest.R
\name{ggttest}
\alias{ggttest}
\title{Student's t-test plot}
\usage{
ggttest(t, colaccept="lightsteelblue1", colreject="grey84", colstat="navyblue")
}
\arguments{
\item{t}{a list result of \code{\link{t.test}} of "htest" class}

\item{colaccept}{color the acceptance area of the test}

\item{colreject}{color for the rejection area of the test}

\item{colstat}{color for the test statistic vline}
}
\description{
Visualise one and/or two sample t-tests on vectors of data.
}
\examples{

t_test <- t.test(sleep$extra ~ sleep$group)
t_test
ggttest(t_test)


t_test2 <- t.test(x = 1:10, y = c(7:20)) 
t_test2
ggttest(t_test2)

}
