% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvartest.R
\name{ggvartest}
\alias{ggvartest}
\title{F test plot}
\usage{
ggvartest(t, colaccept = "lightsteelblue1", 
colreject = "gray84", colstat = "navyblue")
}
\arguments{
\item{t}{a list result of \code{\link{var.test}} of "htest" class}

\item{colaccept}{color the acceptance area of the test, see  \code{\link{colors}}}

\item{colreject}{color for the rejection area of the test}

\item{colstat}{color for the test statistic vline}
}
\description{
Visualise F test to compare two variances
}
\examples{

x <- rnorm(50, mean = 0, sd = 2)
y <- rnorm(30, mean = 1, sd = 1)
var_test <- var.test(x, y)
var_test
ggvartest(var_test)

}
