% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaptopRates.R
\docType{data}
\name{LaptopRates}
\alias{LaptopRates}
\title{LaptopRates}
\format{A data frame with 54 observations of 3 columns:
\describe{
  \item{\code{laptop}}{laptop id, 1-18}
  \item{\code{expert}}{a character of expert1,expert2,expert3 values}
  \item{\code{rating}}{ratings-5 likert scale, 5 indicates a very good rate}
  }}
\usage{
LaptopRates
}
\description{
A data frame that contains sample ratings of 18 laptops, by three experts.
}
