% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivLab.R
\docType{data}
\name{LivLab}
\alias{LivLab}
\title{LivLab}
\format{A data frame with 10 observations of 2 columns:
\describe{
  \item{\code{before}}{score in a neuropsychological test before serious game intervention}
  \item{\code{after}}{score in a neuropsychological test after serious game intervention}
  }}
\source{
<http://aha-livinglabs.com/>
}
\usage{
LivLab
}
\description{
A data frame containing a sample with the results of neuropsychological assessment 
before and after serious game intervention in the living lab, Thess-AHALL 
(Thessaloniki Active and Healthy Aging Living Lab) 
of Medical Physics Laboratory of Aristotle University of Thessaloniki.
}
