% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_anova.R
\docType{data}
\name{m_anova}
\alias{m_anova}
\title{m_anova}
\format{A data frame with 54 observations of 4 columns:
\describe{
  \item{\code{Categories}}{integer representing three categories}
  \item{\code{UrbanAreas}}{coding for urban, suburban and rural areas}
  \item{\code{Month}}{integer representing three months}
  \item{\code{Volume}}{volume in cubic meters}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
m_anova
}
\description{
A data frame with the volume of new stores by category 
for urban, suburban and rural areas.
}
