% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MilkConsumption.R
\docType{data}
\name{MilkConsumption}
\alias{MilkConsumption}
\title{Milk consumption}
\format{A data frame with 13 observations of 3 columns:
\describe{
  \item{\code{Cow.Milk}}{number of cow milk bottles}
  \item{\code{Sheep.Milk}}{number of sheep milk bottles}
  \item{\code{Goat.Milk}}{number of goat milk bottles}
  }}
\source{
The original data are available at 
\href{http://www.statistics.gr/}{Hellenic Statistical Authority}
}
\usage{
MilkConsumption
}
\description{
A data frame containing a sample of the number of cow, 
sheep and goat milk bottles sold.
}
