% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profits_df.R
\docType{data}
\name{profits_df}
\alias{profits_df}
\title{Profits}
\format{
A data frame with 26 companies (rows) of 5 months (columns).
}
\usage{
profits_df
}
\description{
A data frame with the profits of some companies for 5 months,
constructed for teaching purposes.
}
