% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire.R
\docType{data}
\name{questionnaire}
\alias{questionnaire}
\title{Questionnaire Answers}
\format{A data frame with 50 observations of 8 columns:
\describe{
  \item{\code{gender}}{A factor with the student gender}
  \item{\code{writing.hand}}{A factor with the writing hand of the students(left, right)}
  \item{\code{fold.arm}}{A factor with the top hand when the students fold their arms}
  \item{\code{pulse}}{Integer with the pulse rate of students (beats per minute)}
  \item{\code{exercise}}{A factor with the frequency the students exercises (none, some,frequently)}
  \item{\code{smoke}}{A factor with the frequency the students smokes (heavy, regularly, occasionally, never)}
  \item{\code{height}}{Integer with the height of the students (in centimeters)}
  \item{\code{age}}{Integer with the age of the students}
  }}
\usage{
questionnaire
}
\description{
A data frame containing a sample with the answers of students.
}
