% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randexperiment.R
\docType{data}
\name{randexperiment}
\alias{randexperiment}
\title{Random experiment results}
\format{A data frame with 128 observations of 1 column.}
\usage{
randexperiment
}
\description{
A data frame with 128 sample results of a repeated experiment. 
Success is noted with 1 and failure with 0.
}
