% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Salary_Gender.R
\docType{data}
\name{Salary_Gender}
\alias{Salary_Gender}
\title{Female and male salaries}
\format{A data frame with 100 observations of 2 columns:
\describe{
  \item{\code{Male_sal}}{male salaries}
  \item{\code{Female_sal}}{female salaries}
  }}
\usage{
Salary_Gender
}
\description{
A sample data frame with female and male salaries of a company.
}
